+++ src/SeerGdbWidget.h
@@ -107,15 +107,6 @@ class SeerGdbWidget : public QWidget, protected Ui::SeerGdbWidgetForm {
         void                                setGdbRecordDirection               (const QString& direction);
         QString                             gdbRecordDirection                  () const;
 
-        void                                setDprintfStyle                     (const QString& style);
-        QString                             dprintfStyle                        () const;
-
-        void                                setDprintfFunction                  (const QString& function);
-        QString                             dprintfFunction                     () const;
-
-        void                                setDprintfChannel                   (const QString& channel);
-        QString                             dprintfChannel                      () const;
-
         void                                setConsoleMode                      (const QString& mode);
         QString                             consoleMode                         () const;
 
@@ -202,9 +193,6 @@ class SeerGdbWidget : public QWidget, protected Ui::SeerGdbWidgetForm {
         void                                writeSettings                       ();
         void                                readSettings                        ();
 
-        // Printpoints
-        void                                resetDprintf                        ();
-
     public slots:
         void                                handleLogsTabMoved                  (int from, int to);
         void                                handleLogsTabChanged                (int index);
@@ -276,6 +264,7 @@ class SeerGdbWidget : public QWidget, protected Ui::SeerGdbWidgetForm {
         void                                handleGdbBreakpointInsert           (QString breakpoint);
         void                                handleGdbBreakpointCondition        (QString breakpoint, QString condition);
         void                                handleGdbBreakpointIgnore           (QString breakpoint, QString count);
+        void                                handleGdbBreakpointCommand          (QString breakpoint, QString commands);
         void                                handleGdbBreakpointCommands         (QString breakpoint, QStringList commands);
         void                                handleGdbBreakpointReload           (QStringList breakpointsText);
         void                                handleGdbWatchpointReload           (QStringList watchpointsText);
@@ -291,7 +280,7 @@ class SeerGdbWidget : public QWidget, protected Ui::SeerGdbWidgetForm {
         void                                handleGdbPrintpointDelete           (QString breakpoints);
         void                                handleGdbPrintpointEnable           (QString breakpoints);
         void                                handleGdbPrintpointDisable          (QString breakpoints);
-        void                                handleGdbPrintpointInsert           (QString printpoint);
+        void                                handleGdbPrintpointInsert           (QString type, QString function, QString channel, QString parameters);
         void                                handleGdbThreadListFrames           ();
         void                                handleGdbThreadListIds              ();
         void                                handleGdbThreadListGroups           ();
@@ -382,9 +371,6 @@ class SeerGdbWidget : public QWidget, protected Ui::SeerGdbWidgetForm {
         bool                                _gdbEnablePrettyPrinting;
         QString                             _gdbRecordMode;
         QString                             _gdbRecordDirection;
-        QString                             _dprintfStyle;
-        QString                             _dprintfFunction;
-        QString                             _dprintfChannel;
         bool                                _assemblyShowAssemblyTabOnStartup;
         QString                             _assemblyDisassemblyFlavor;
         QString                             _assemblySymbolDemangling;