+++ CHANGELOG.md
@@ -25,6 +25,10 @@
 * Source all files in ~/.config/seergdb/scripts/ on startup.
 * Fixed regression handling ignoring of system header files.
 * Fixed regression when adding commands to a breakpoint.
+* Refactored/improved Printpoints. Added 'help' to the
+  Printpoint create dialog.
+* Fixed regression when display variable value when hovering over
+  the text in the editor window.
 
 ## [2.4] - 2024-03-18
 * Changed main icon to a more license friendly one.